Update ArrayPattern and java bindings
